@@ -40,8 +40,10 @@
     "babel-cli": "^6.24.1",
     "babel-loader": "^7.1.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
     "body-parser": "^1.17.2",
     "compression": "^1.6.2",
     "cookie-parser": "^1.4.3",
@@ -59,9 +61,10 @@
     "react-redux": "^5.0.5",
     "redux": "^3.7.1",
     "redux-logger": "^3.0.6",
+    "styled-components": "^2.1.1",
     "uglifyjs-webpack-plugin": "^0.4.6",
     "universal-router": "^3.2.0",
-    "webpack": "^3.0.0",
+    "webpack": "^3.3.0",
     "webpack-merge": "^4.1.0"
   },
   "devDependencies": {
