@@ -1,14 +1,14 @@
-const webpack = require('webpack');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+import webpack from 'webpack'
+import UglifyJsPlugin from 'uglifyjs-webpack-plugin'
 
-module.exports = {
+export default {
   devtool: 'sourcemap',
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
-        NODE_ENV: JSON.stringify('production'),
-      },
+        NODE_ENV: JSON.stringify('production')
+      }
     }),
-    new UglifyJsPlugin(),
-  ],
-};
+    new UglifyJsPlugin()
+  ]
+}
