@@ -1,12 +1,8 @@
-const postcssImport = require('postcss-import');
-const postcssSorting = require('postcss-sorting');
-const csswring = require('csswring');
+const postcssImport = require('postcss-import')
+const postcssSorting = require('postcss-sorting')
+const csswring = require('csswring')
 
-module.exports = (ctx) => ({
+module.exports = ctx => ({
   map: ctx.options.map,
-  plugins: [
-    postcssImport(),
-    postcssSorting(),
-    csswring()
-  ]
-});
+  plugins: [postcssImport(), postcssSorting(), csswring()]
+})
